test(fuzz): repair fuzz workspaces, gate them in CI, fix proto-28 filter prefix parity (#6997)
* test(fuzz): repair broken fuzz targets and compile-gate fuzz workspaces

The fuzz workspaces are excluded from the root workspace so libfuzzer-sys
stays out of ordinary builds, which also hid compile breakage from every
PR gate. Four targets in fuzz/ no longer compiled against current crate
APIs and the weekly Fuzz Coverage Smoke has never passed since its
introduction on 2026-05-18.

- auth_response: port to the digest-parameterized verifier API and add a
  soundness/completeness oracle per negotiated digest.
- filter_list_wire: import read_filter_list from its public re-export and
  add a write/read round-trip oracle.
- parallel_receive_delta_adversarial: FileNdx conversions take u32.
- rsyncd_conf: removed; its parser was deleted in #6672 and the unified
  replacement is crate-private, so the target is unreachable.
- differential_flist: apply fuzzed mtime/uid/gid so they reach the wire
  encoder (fixes dead_code under -D warnings).
- cargo fmt across fuzz targets; refresh stale protocol-fuzz lockfile.

CI: new fuzz-workspace-check job runs tools/ci/check_fuzz_workspaces.sh -
cargo check + cargo fmt --check per fuzz workspace on stable, so the PR
that introduces API drift fails instead of the next overnight fuzz run.

fuzz-coverage.yml: install cargo-fuzz as a prebuilt binary (source
install no longer builds under the pinned stable toolchain) and fix an
unbound host_triple under set -u in the coverage step.

* fix(protocol): match upstream optional old-style filter prefixes

The strengthened filter_list_wire re-encode oracle immediately found two
proto < 29 reader divergences from upstream parse_rule_tok():

- A bare pattern was rejected with 'invalid old-style filter prefix'.
  Upstream treats '- ' and '+ ' as optional under XFLG_OLD_PREFIXES
  (exclude.c:1125-1133) and its sender emits plain excludes with no
  prefix at these protocols (get_rule_prefix, exclude.c:1542), so every
  upstream 2.x-era exclude push was rejected by our receiver.
- An empty pattern after a prefix was accepted. Upstream exits
  RERR_SYNTAX with 'unexpected end of filter rule' (exclude.c:1324-1327).

The wire_parser_rejects_dir_merge_at_protocol_28 test encoded the first
divergence (upstream's receiver parses a ':'-prefixed payload as a bare
exclude; only the sender refuses to emit it) and is updated to assert the
upstream behaviour. New unit tests cover the bare-exclude wire bytes and
the '!text' exclude form.

The fuzz oracle now asserts encode/decode idempotence rather than literal
equality, since upstream normalizes some received forms (P -> -r).

* ci: honour --locked in the fuzz workspace guard

check-locked-flags gates every cargo check in tools/ci. Commit the
fuzz/ and crates/filters/fuzz lockfiles (crates/protocol/fuzz was
already tracked) and re-sync each lockfile offline before the --locked
check, mirroring the lint job, so workspace version bumps do not wedge
the guard.

+++ fuzz/fuzz_targets/parallel_receive_delta_adversarial.rs
@@ -99,8 +99,7 @@ fn parse_input(data: &[u8]) -> Option<(u8, Vec<Record>)> {
         return None;
     }
     let file_count = (data[0] % MAX_FILES) + 1;
-    let payload_log2 = MIN_PAYLOAD_LOG2
-        + (data[1] % (MAX_PAYLOAD_LOG2 - MIN_PAYLOAD_LOG2 + 1));
+    let payload_log2 = MIN_PAYLOAD_LOG2 + (data[1] % (MAX_PAYLOAD_LOG2 - MIN_PAYLOAD_LOG2 + 1));
     let payload_size = 1usize << payload_log2;
     let record_size = 1 + 2 + payload_size;
 
@@ -223,13 +222,13 @@ fn parallel_destination_hashes(
     for file in 0..file_count {
         let (sink, buf) = VecSink::new();
         applier
-            .register_file(u64::from(file), Box::new(sink) as Box<dyn Write + Send>)
+            .register_file(u32::from(file), Box::new(sink) as Box<dyn Write + Send>)
             .ok()?;
         buffers.insert(file, buf);
     }
     for r in records {
         let chunk = DeltaChunk::literal(
-            u64::from(r.file_ndx),
+            u32::from(r.file_ndx),
             u64::from(r.chunk_sequence),
             r.payload.clone(),
         );
@@ -263,7 +262,8 @@ fuzz_target!(|data: &[u8]| {
         return;
     };
     assert_eq!(
-        actual, expected,
+        actual,
+        expected,
         "parallel receive-delta SHA-256 divergence \
          (file_count={file_count}, records={})",
         records.len(),
